using ft_strlcpy in tokenizer

+++ include/utils.h
@@ -7,7 +7,7 @@
 #include "ft_error.h"
 
 size_t	ft_strlen(char *s);						// string size
-char	*ft_strlcpy(char *dest, const char *src, int n);
+char	*ft_strlcpy(char *dest, const char *src, size_t n);
 void	*ft_calloc(size_t count, size_t size);
 void	*ft_memset(void *dest, int c, size_t len);
 char	*ft_strcat(char *dest, char *src);
