+++ src/utils/ft_strlcpy.c
@@ -1,4 +1,18 @@
-char	*ft_strlcpy(char *dest, const char *src, int n)
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_strlcpy.c                                       :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: sabdelra <sabdelra@student.42abudhabi.a    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/06/09 12:46:12 by sabdelra          #+#    #+#             */
+/*   Updated: 2023/06/09 12:47:01 by sabdelra         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include <stdlib.h>
+
+char	*ft_strlcpy(char *dest, const char *src, size_t n)
 {
 	int	i;
